requirements.txt
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.24
fastapi==0.104.1
uvicorn==0.24.0
pydantic
python-dotenv==1.0.0
pyyaml==6.0.1
numpy==1.24.3
pandas==2.0.3
ollama

